Code Review Cleanup. Files now all pass flake8 tests
