+++ tests/ext/sqlalchemy/test_query.py
@@ -6,12 +6,12 @@ from aws_xray_sdk.core import xray_recorder
 from aws_xray_sdk.core.context import Context
 from aws_xray_sdk.ext.sqlalchemy.query import XRaySessionMaker
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import *
+from sqlalchemy import create_engine, Column, Integer, String
 # from sqlalchemy.orm import sessionmaker
 
 
-
 Base = declarative_base()
+
 
 class User(Base):
         __tablename__ = 'users'
@@ -21,7 +21,7 @@ class User(Base):
         fullname = Column(String)
         password = Column(String)
 
- 
+
 def _search_entity(entity, name):
     """Helper function to that recursivly looks at subentities
     Returns a serialized entity that matches the name given or None"""
@@ -32,19 +32,22 @@ def _search_entity(entity, name):
         else:
             if "subsegments" in entity:
                 for s in entity['subsegments']:
-                    result = _search_entity(s,name)
-                    if result != None:
+                    result = _search_entity(s, name)
+                    if result is not None:
                         return result
     return None
+
+
 def find_sub(segment, name):
     """Helper function to find a subsegment by name in the entity tree"""
     segment = jsonpickle.encode(segment, unpicklable=False)
     segment = json.loads(segment)
     for entity in segment['subsegments']:
         result = _search_entity(entity, name)
-        if result != None:
+        if result is not None:
             return result
     return None
+
 
 @pytest.fixture()
 def session():
@@ -59,7 +62,8 @@ def session():
     yield session
     xray_recorder.end_segment()
     xray_recorder.clear_trace_entities()
-    
+
+
 def test_all(capsys, session):
     """ Test calling all() on get all records.
     Verify we run the query and return the SQL as metdata"""
@@ -74,9 +78,7 @@ def test_add(capsys, session):
     """ Test calling add() on insert a row.
     Verify we that we capture trace for the add"""
     # with capsys.disabled():
-    john = User(name='John', fullname = "John Doe", password="password")
+    john = User(name='John', fullname="John Doe", password="password")
     session.add(john)
     sub = find_sub(xray_recorder.current_segment(), 'sqlalchemy.orm.session.add')
     assert sub['name'] == 'sqlalchemy.orm.session.add'
-        
-
